Add files via upload
Added an example file that looks for convergence in a set of parallel analyses of the same layer.analyzer model. the example is that of a causal connection model for bivalve extinction rate going to brachipod extinction rate. The model has 2 layers for both time series. After analysing this N times, plots and Gelman diagnostics are shown. The code then concentrates on what the Gelman diagnostics regards as the worst parameter, plotting the chains and performing auto-correlation analysis on the full array of results.